requirements.txt
fastapi==0.86.0
pydantic==1.10.2
uvicorn[standard]==0.19.0
Twisted==22.10.0
autobahn==22.7.1
pyopenssl==22.1.0
service_identity==21.1.0
ujson==5.5.0
ruamel.yaml==0.17.21
click==8.1.3
requests==2.28.1
setuptools==65.5.1
jsons==1.6.3
